added task 0 of doubly linked list

+++ 0x17-doubly_linked_lists/0-print_dlistint.c
@@ -1,4 +1,3 @@
-#include <stdlib.h>
 #include <stdio.h>
 #include "lists.h"
 
@@ -10,7 +9,7 @@
  */
 size_t print_dlistint(const dlistint_t *h)
 {
-    	int count = 0;
+    	size_t count = 0;
 	
 	if (!h)
 		return (count);
